requirements.txt
streamlit
replicate
snowflake-ml-python
